fix: 기존의 이미지 관련 rscService -> Service 훅으로 변경 및 호출 단순화 (#146)
