+++ src/shared/hooks/services/UserService.tsx
@@ -1,6 +1,7 @@
 import { AxiosResponse } from "axios";
 
-import { API, useUserStore } from "@/shared";
+import { API, FORMAPI, useUserStore } from "@/shared";
+import { createFormData } from "@/utils";
 
 export const UserService = () => {
   const URI = "/api/v1/users";
@@ -85,6 +86,60 @@ export const UserService = () => {
     }
   };
 
+  const submitAdmissionsApplication = async (
+    data: User.AdmissionCreateRequestDto, // FileList 타입 사용
+  ): Promise<any> => {
+    try {
+      const payload = {
+        ...data,
+        images: undefined, 
+      };
+
+      const formData = createFormData(
+        payload,
+        "userAdmissionCreateRequestDto", 
+        data.attachImage ? Array.from(data.attachImage) : [], 
+        "userAdmissionAttachImageList" 
+      );
+  
+      const response = await FORMAPI.post(
+        "/api/v1/users/admissions/apply",
+        formData,
+      );
+  
+      return response.data; //
+    } catch (error) {
+      throw error;
+    }};
+
+
+  const updateInfo = async (
+    data: User.userUpdateDto, // FileList 타입 사용
+  ): Promise<any> => {
+    try {
+      const payload = {
+        ...data,
+        images: undefined, 
+      };
+
+      const formData = createFormData(
+        payload,
+        "userUpdateDto", 
+        data.profileImage ? [data.profileImage] : [], 
+        "profileImage" 
+      );
+  
+      const response = await FORMAPI.put(
+        "/api/v1/users",
+        formData,
+      );
+  
+      return response.data; //
+    } catch (error) {
+      throw error;
+    }};
+
+
   return {
     getMe,
     getMyInfo,
@@ -95,5 +150,7 @@ export const UserService = () => {
     checkCurrentAcademicStatus,
     checkIsAcademicRecordSubmitted,
     allowUser,
+    submitAdmissionsApplication,
+    updateInfo,
   };
 };
